pyproject.toml
[tool.poetry]
name = "nf-ai-agent-chatbot"
version = "0.1.0"
description = "A Q&A chatbot on planning applications in London"
authors = ["Kevin Lam <kevinlam235@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<4.0"
openai = "^1.0.0"
python-dotenv = "^1.0.0"
pandas = "1.5.3"
langchain = "^0.3.22"
langchain-openai = "^0.3.12"
langchain-core = "^0.3.50"
langchain-experimental = "^0.0.36"
faiss-cpu = "^1.7.0"
protobuf = "==3.20.*"
streamlit = "^1.31.0"
pymupdf = "^1.23.0"
google-api-python-client = "^2.0.0"
tqdm = "^4.66.0"
numpy = "1.26.4"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
black = "^23.3.0"
flake8 = "^6.0.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88

[tool.isort]
profile = "black"
line_length = 88